+++ pkg/config/config.go
@@ -167,7 +167,10 @@ func (c *Config) Merge(config Config) {
 	}
 
 	c.mergeDisabled(config.Disable)
-	c.Logger = config.Logger
+	c.Logger = logger.Logger{
+		Verbosee: c.Verbose || config.Verbose,
+		Debugg:   c.Debug || config.Debug,
+		NoColor:  c.NoColor || config.NoColor}
 }
 
 // mergeDisabled merges the disabled checks into the config